# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "知行 ZhiXing Personal Agent",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/search": {
      "post": {
        "summary": "Search",
        "operationId": "search_api_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mail/accounts/{ident}/reindex": {
      "post": {
        "summary": "Reindex",
        "operationId": "reindex_api_mail_accounts__ident__reindex_post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mail/drafts": {
      "get": {
        "summary": "Drafts",
        "operationId": "drafts_api_mail_drafts_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Draft",
        "operationId": "create_draft_api_mail_drafts_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mail/drafts/{ident}": {
      "post": {
        "summary": "Update Draft",
        "operationId": "update_draft_api_mail_drafts__ident__post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mail/drafts/{ident}/send": {
      "post": {
        "summary": "Confirm Send",
        "description": "A user's final click authorizes exactly this frozen draft version.",
        "operationId": "confirm_send_api_mail_drafts__ident__send_post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Version"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assistant/sessions": {
      "get": {
        "summary": "Sessions",
        "operationId": "sessions_api_assistant_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Session",
        "operationId": "session_api_assistant_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assistant/sessions/{ident}/turns": {
      "post": {
        "summary": "Turn",
        "operationId": "turn_api_assistant_sessions__ident__turns_post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assistant/sessions/{ident}": {
      "get": {
        "summary": "Session Detail",
        "operationId": "session_detail_api_assistant_sessions__ident__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assistant/turns/{ident}": {
      "get": {
        "summary": "Turn Detail",
        "operationId": "turn_detail_api_assistant_turns__ident__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assistant/turns/{ident}/cancel": {
      "post": {
        "summary": "Cancel Turn",
        "operationId": "cancel_turn_api_assistant_turns__ident__cancel_post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DraftInput": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "new",
              "reply",
              "reply_all"
            ],
            "title": "Mode",
            "default": "new"
          },
          "to": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "To"
          },
          "cc": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Cc"
          },
          "subject": {
            "type": "string",
            "maxLength": 500,
            "title": "Subject",
            "default": ""
          },
          "content": {
            "type": "string",
            "maxLength": 50000,
            "title": "Content",
            "default": ""
          },
          "version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Version",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "account_id"
        ],
        "title": "DraftInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchRequest": {
        "properties": {
          "account_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Account Ids"
          },
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "sender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender"
          },
          "mode": {
            "type": "string",
            "enum": [
              "hybrid",
              "keyword",
              "vector",
              "fusion"
            ],
            "title": "Mode",
            "default": "hybrid"
          }
        },
        "type": "object",
        "required": [
          "account_ids",
          "query"
        ],
        "title": "SearchRequest"
      },
      "SessionInput": {
        "properties": {
          "account_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Account Ids"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          }
        },
        "type": "object",
        "required": [
          "account_ids"
        ],
        "title": "SessionInput"
      },
...
      "TurnInput": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TurnInput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Version": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "version"
        ],
        "title": "Version"
      }
    }
  }
}